requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
Werkzeug
openpyxl
reportlab